<feat>(MobX): Add Navigation store
