--- packages/client-web/src/store/UiStore.ts
@@ -1,11 +1,0 @@
-import RootStore from './RootStore';
-
-class UiStore {
-  rootStore: RootStore;
-
-  constructor(rootStore: RootStore) {
-    this.rootStore = rootStore;
-  }
-}
-
-export default UiStore;
