+++ packages/client-web/src/store/UiStore/index.ts
@@ -1,0 +1,15 @@
+import RootStore from '../RootStore';
+import { observable, action } from 'mobx';
+import Navigation from './Navigation';
+
+class UiStore {
+  rootStore: RootStore;
+  navigation: Navigation;
+
+  constructor(rootStore: RootStore) {
+    this.rootStore = rootStore;
+    this.navigation = new Navigation(rootStore);
+  }
+}
+
+export default UiStore;
